Refactored for storage client
