+++ medicines/doc-index-updater/src/storage_client/mod.rs
@@ -1,79 +1,18 @@
-use async_trait::async_trait;
-use azure_sdk_core::{errors::AzureError, prelude::*, DeleteSnapshotsMethod};
-use azure_sdk_storage_blob::Blob;
+use azure_sdk_core::errors::AzureError;
 use azure_sdk_storage_core::prelude::Client;
 
-pub struct BlobClient {
-    pub azure_client: Client,
-}
+pub use client::BlobClient;
+pub use client::StorageClient;
+pub use delete::DeleteBlob;
+pub use get::GetBlob;
 
-impl BlobClient {
-    pub fn new(azure_client: Client) -> BlobClient {
-        BlobClient { azure_client }
-    }
-}
-
-#[async_trait]
-pub trait DeleteBlob {
-    async fn delete_blob(
-        &mut self,
-        container_name: &str,
-        blob_name: &str,
-    ) -> Result<(), AzureError>;
-}
-
-#[async_trait]
-impl DeleteBlob for BlobClient {
-    async fn delete_blob(
-        &mut self,
-        container_name: &str,
-        blob_name: &str,
-    ) -> Result<(), AzureError> {
-        self.azure_client
-            .delete_blob()
-            .with_container_name(&container_name)
-            .with_blob_name(&blob_name)
-            .with_delete_snapshots_method(DeleteSnapshotsMethod::Include)
-            .finalize()
-            .await?;
-        Ok(())
-    }
-}
-
-#[async_trait]
-pub trait GetBlob {
-    async fn get_blob(
-        &mut self,
-        container_name: &str,
-        blob_name: &str,
-    ) -> Result<BlobResponse, AzureError>;
-}
+mod client;
+mod delete;
+mod get;
 
 pub struct BlobResponse {
     pub blob_name: String,
     pub data: Vec<u8>,
-}
-
-#[async_trait]
-impl GetBlob for BlobClient {
-    async fn get_blob(
-        &mut self,
-        container_name: &str,
-        blob_name: &str,
-    ) -> Result<BlobResponse, AzureError> {
-        let blob = self
-            .azure_client
-            .get_blob()
-            .with_container_name(&container_name)
-            .with_blob_name(&blob_name)
-            .finalize()
-            .await?;
-
-        Ok(BlobResponse {
-            blob_name: blob.blob.name,
-            data: blob.data,
-        })
-    }
 }
 
 pub fn factory() -> Result<BlobClient, AzureError> {
